docs-package: don't initialize the `readFilesProcessor`
The current setup for angular.io does not use the `readFilesProcessor`
so the `docs-package` doesn't need to configure it.

(For example, all the guide docs are written in straight Jade and are not
processed by dgeni).

In fact the `angular.io-package` explicitly disables this processor.

Given this, setting the basePath for the `readFilesProcessor` in the
`angular.io-package` is not really a hack, since we do need it for the
`writeFilesProcessor`.