chore: optimize gulp watchers
Reduce amount of watchers for `gulp serve-and-sync-devguide` by roughly 50% (~4900-ish to ~2500).

This should help with the 100% CPU usage #1082, on my machine it went from ~90% to ~50%.

At the end of the day we're still watching an insane amount of files, so it is a demanding process. Maybe using better watcher libs would also help.