Cookbook for a component item template.
This cookbook demonstrates how to provide a component with an external
`TempalteRef` that the component can then use to render aspects of its own
internal template. An example of this, as explored in the cookbook, is
providing a select-menu component with an item template that the component
can use to render the options in its dropdown.