docs(dependency-injection): Clarify how TypeScript generates decorator metadata
I think it's important to mention the compiler option. I wanted to see how this magic works and what it looks like in a compiled javascript and tried to make a simple test in TypeScript playground. I got nothing and it took me some time to figure out that a compiler option is involved in this.