Custom exception handler service cookbook.
This cookbook demonstrates how to override the core `ExceptionHandler` class
with a custom service class that both logs to the console as well as sending
errors to a robust error-logginer service. This error-logging service can,
in turn, send errors to the server; or, to Software-as-a-Service (SaaS)
products like New Relic, TrackJS, and Raygun.