docs(toh-5): review and update/resync Dart
**NOTE: run `gulp add-example-boilerplate` after pulling in the
commit.**

This is preparatory work for #2035.
As part of the the chapter review, the Dart .jade was enhanced to use
Jade extends (#2018).
By the same token it contributed to a post-RC5 resync (#2077). Other
key changes:

Dart and TS code:
- Eliminated `styles.1.css` in favor of docregions in `styles.css`.
- `docregion` tags renamed in a few places.
- **No other code changes**.

TS prose
- Fixed: misnamed variable `routing` -> `appRoutes`.
- All other changes are **minor copy edits**, or changes to support
Dart via Jade extends.

Diff of generated HTML for TS chapter was inspected to ensure only
minor copy edits prevailed (i.e., that the support for Jade extends had
no impact on the generated HTML).