example(toh-4,5): getHeroesSlowly() to return getHeroes() (#2152)
* example(dart/toh-4,5): getHeroesSlowly() to return getHeroes()

Have `getHeroesSlowly()` delay and then return the value of
`getHeroes()`. This makes it easier for user’s performing the tutorial
to keep this slower method as they evolve toh-5 into toh-6.

* example(ts/toh-4,5): getHeroesSlowly() to return getHeroes()

Have `getHeroesSlowly()` delay and then return the value of
`getHeroes()`. This makes it easier for user’s performing the tutorial
to keep this slower method as they evolve toh-5 into toh-6.