chore(examples): don't gitignore _examples/tsconfig.json
Tweak to `.gitignore` so that only the `tsconfig.json` files in
subdirectories are ignored, not the one directly in this `_examples`
folder.

Also, add a package.json script definition for `http-server` for more
convenient access (w/o invoking tsc as is forced in the script
definition given in `_boilerplate/package.json`.