docs(template-syntax): Rectify the concept of 'context'
A template reference variable (Component, Directive, ElementRef or TemplateRef instance) has nothing to do with a context. If it's a Component instance, it may be a context of another template, but not relevant in this template, nor alternative; if it's a TemplateRef instance, then it would always depends on an external context object to be passed in as the view model and needs template input variable to define what to get from that context; if it's a Directive or ElementRef instance, there is nothing todo with context at all.