docs: update CLI quickstart to rc.0
RC0 hasn't been released yet but will soon.

All shown boilerplate is copy paste of a new CLI generated project using the rc.0 blueprints.

Formatting might be a bit different from the norm in the angular.io repo but it should be left as is, since the shown files should match a generated project.

For angular.io tests to pass this PR is still using the `@angular/cli@1.0.0-beta.32.3` npm version however.

Prior to merging it should be updated to `1.0.0-rc.0` in these two places:

- `public/docs/_examples/package.json`
- `public/docs/_examples/_boilerplate/package.json`